change Datatype to FLOAT(10, 2)

+++ back/models/order-model.js
@@ -1,7 +1,7 @@
 const orderModel = (sequelize, DataTypes) => {
     const Order = sequelize.define("Order", {
         totalPrice: {
-            type: DataTypes.DECIMAL,
+            type: DataTypes.FLOAT(10,2),
             allowNull: false
         }
     })
